pyproject.toml
[tool.poetry]
name = "bedrock-server"
version = "0.1.0"
description = ""
authors = ["dev-laww <lawrenceallen.tech@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
mcstatus = "^11.1.1"
python-dotenv = "^1.1.0"
rich = "^14.0.0"
readchar = "^4.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
